+++ src/extensions/renderer-api/kube-object-status.ts
@@ -1,0 +1,11 @@
+export type KubeObjectStatus = {
+  level: KubeObjectStatusLevel;
+  text: string;
+  timestamp?: string;
+}
+
+export enum KubeObjectStatusLevel {
+  INFO = 1,
+  WARNING = 2,
+  CRITICAL = 3
+}
